pagina/apps/Registro/serializers.py
from rest_framework import serializers
from .models import Anime, Manga

class AnimeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Anime
        fields = '__all__'
        
class MangaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Manga
        fields = '__all__'
