requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.3
django-pwa==1.0.10
djangorestframework==3.12.2
idna==2.10
oauthlib==3.1.0
Pillow==8.0.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
